% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVarimp.R
\name{aggregateVarimp}
\alias{aggregateVarimp}
\title{Aggregate Variable Importance Estimates}
\usage{
aggregateVarimp(
  vimp,
  aggregate = c("mean", "median"),
  scale = c("absolute", "relative.baseline"),
  omit.na = TRUE,
  scale.by = 1
)
}
\arguments{
\item{vimp}{Variable importance estimate from a SEM forest.}

\item{aggregate}{Character. Either 'mean' or 'median' as function to aggregate estimates over a forest}

\item{scale}{Character. Either 'absolute' or 'relative'.}

\item{omit.na}{Boolean. By default TRUE, which ignores NA estimates when aggregating. Otherwise they are interpreted as zero.}

\item{scale.by}{Integer. By default 1. Scaling parameter.}
}
\description{
This function aggregates variable importance estimates over
trees. It is a helper function used when print() is called
on a variable importance estimate from a SEM forest.
}
