% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodiversity.R
\name{biodiversity}
\alias{biodiversity}
\title{Quantify bio diversity of a SEM Forest}
\usage{
biodiversity(x, aggregate.fun = median)
}
\arguments{
\item{x}{A \code{\link{semforest}} object}

\item{aggregate.fun}{Takes a function to apply to the vector of pairwise
diversities. By default, this is the median.}
}
\description{
A function to calculate biodiversity of a \code{\link{semforest}} object.
}
\author{
Andreas M. Brandmaier
}
\keyword{biodiversity}
\keyword{semforest}
