% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{boruta}
\alias{boruta}
\title{BORUTA algorithm for SEM trees}
\usage{
boruta(
  model,
  data,
  control = NULL,
  predictors = NULL,
  percentile_threshold = 100,
  rounds = 1,
  ...
)
}
\arguments{
\item{model}{A template model specification from \code{\link{OpenMx}} using
the \code{\link{mxModel}} function (or a \code{\link[lavaan]{lavaan}} model
using the \code{\link[lavaan]{lavaan}} function with option fit=FALSE).
Model must be syntactically correct within the framework chosen, and
converge to a solution.}

\item{data}{Data.frame used in the model creation using
\code{\link{mxModel}} or \code{\link[lavaan]{lavaan}} are input here. Order
of modeled variables and predictors is not important when providing a
dataset to \code{semtree}.}

\item{control}{\code{\link{semtree}} model specifications from
\code{\link{semtree.control}} are input here. Any changes from the default
setting can be specified here.}

\item{percentile_threshold}{Numeric.}

\item{rounds}{Numeric. Number of rounds of the BORUTA algorithm.}
}
\description{
BORUTA algorithm for SEM trees
}
