% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePval_maxLR.R
\name{computePval_maxLR}
\alias{computePval_maxLR}
\title{Wrapper function for computing the maxLR corrected p value
from strucchange}
\usage{
computePval_maxLR(maxLR, q, covariate, from, to, nrep)
}
\arguments{
\item{from, }{to: numeric from interval (0, 1) specifying start and end of trimmed}

\item{maxLR:}{maximum of the LR test statistics}

\item{q:}{number of free SEM parameters / degrees of freedom}

\item{covariate:}{covariate under evaluation. This is important to get the level of}

\item{nrep:}{numeric. Number of replications used for simulating from the asymptotic}
}
\value{
Numeric. p value for maximally selected LR statistic
}
\description{
Wrapper function for computing the maxLR corrected p value
from strucchange
}
\author{
Manuel Arnold
}
