% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPredictors.R
\name{countPredictors}
\alias{countPredictors}
\alias{countPredictors.semtree}
\alias{countPredictors.semforest}
\title{Count predictors}
\usage{
countPredictors(x, ...)

\method{countPredictors}{semtree}(x, ...)

\method{countPredictors}{semforest}(x, ...)
}
\arguments{
\item{x}{An object representing either a SEM tree or a forest.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A result depending on the method.
}
\description{
This function is a generic to count the number of occurences of
predictors in either a 
\code{"semtree"} or a \code{"semforest"}. Note that this must not
be confused with the importance of those predictors. To estimate 
importance, rather use permutation-based variable importance with
\code{varimp}.
}
