\name{Distances}
\alias{hellinger}
\alias{klsym}
\title{Distances}
\description{Divergence measures for multivariate normal distributions as used in the diversityMatrix function.}
\usage{
 klsym(mu1, cov1, mu2, cov2)
 hellinger(mu1, cov1, mu2, cov2)
}
\arguments{
\item{mu1}{Mean vector}
\item{mu2}{Mean vector}
\item{cov1}{Covariance matrix}
\item{cov2}{Covariance matrix}
}