% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.semforest.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.semforest}
\title{Average Deviance of a Dataset given a Forest}
\usage{
evaluate(x, data = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{\link{semforest}} object}

\item{data}{A data.frame}

\item{\dots}{No extra parameters yet.}
}
\value{
Average deviance
}
\description{
Evaluates the average deviance (-2LL) of a dataset given a forest.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\seealso{
\code{\link{evaluateDataLikelihood}}, \code{\link{semtree}},
\code{\link{semforest}}
}
\author{
Andreas M. Brandmaier
}
