% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSubmodels.R
\name{fitSubmodels}
\alias{fitSubmodels}
\title{Fit multigroup model for evaluating a candidate split}
\usage{
fitSubmodels(
  model,
  subset1,
  subset2,
  control,
  invariance = NULL,
  return.models = FALSE
)
}
\arguments{
\item{model}{A model specification that is used as template for each of the two groups}

\item{subset1}{Dataset for the first group model}

\item{subset2}{Dataset for the second group model}

\item{control}{a \code{semtree.control} object}

\item{invariance}{fit models with invariant parameters if given. NULL otherwise (default).}

\item{return.models}{boolean. Return the fitted models
returns NA if fit fails}
}
\description{
Fit multigroup model for evaluating a candidate split
}
