% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeight.R
\name{getDepth}
\alias{getDepth}
\title{Get the depth (or, height) a tree.}
\usage{
getDepth(tree)
}
\arguments{
\item{tree}{A \code{\link{semtree}} object}
}
\description{
Returns the length of the longest path from a root node to a leaf node.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier
}
