% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeById.R
\name{getNodeById}
\alias{getNodeById}
\title{Get Node By Id}
\usage{
getNodeById(tree, id)
}
\arguments{
\item{tree}{A SEM Tree object.}

\item{id}{Numeric. A Node id.}
}
\description{
Return a node matching a given node ID
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
