% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumNodes.R
\name{getNumNodes}
\alias{getNumNodes}
\title{Tree Size}
\usage{
getNumNodes(tree)
}
\arguments{
\item{tree}{A SEM tree object.}
}
\description{
Counts the number of nodes in a tree.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
