% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParContrib.R
\name{getParContrib}
\alias{getParContrib}
\title{Return table with parameter contributions to the score-based test
statistics}
\usage{
getParContrib(tree, standardized = FALSE)
}
\arguments{
\item{tree}{a semtree object.}

\item{standardized}{logical value indicating whether standardized or raw
(default) contributions are to be returned.}
}
\value{
A matrix containing the parameter contributions to the score-based
test statistics. The matrix has \eqn{n} rows and \eqn{k} columns, where
\eqn{n} is the number of non-leaf nodes of the tree and \eqn{k} is the number
of model parameters. The rows are named by the node IDs as given
\code{getNodeById} and the columns are named as in \code{coef}.
}
\description{
Returns a table with the contributions of the model parameters
to the score-based test statistics.
}
\author{
Manuel Arnold
}
