% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandDiff.R
\name{getStandDiff}
\alias{getStandDiff}
\title{Return table with standardized differences of post-split parameters}
\usage{
getStandDiff(tree)
}
\arguments{
\item{tree}{a semtree object.}
}
\value{
A matrix containing the standardized parameter differences. The
matrix has \eqn{n} rows and \eqn{k} columns, where \eqn{n} is the number of
non-leaf nodes of the tree and \eqn{k} is the number of model parameters. The
rows are named by the node IDs as given \code{getNodeById} and the columns
are named as in \code{coef}.
}
\description{
Returns a table with the standardized differences of post-split
parameters. Standardized differences are computed by calculating the
differences of the post-split nodes devided by the pooled standard errors.
}
\author{
Manuel Arnold
}
