% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerminalNodes.R
\name{getTerminalNodes}
\alias{getTerminalNodes}
\title{Returns all leafs of a tree}
\usage{
getTerminalNodes(tree)
}
\arguments{
\item{tree}{A semtree object.}
}
\description{
Returns all leafs (=terminal nodes) of a tree.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
