% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeaf.R
\name{isLeaf}
\alias{isLeaf}
\title{Test whether a semtree object is a leaf.}
\usage{
isLeaf(tree)
}
\arguments{
\item{tree}{A \code{\link{semtree}} object}
}
\description{
Tests whether a semtree object is a leaf. Returns TRUE or FALSE.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier
}
