% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl}
\alias{kl}
\alias{hellinger}
\alias{klsym}
\title{Distances}
\usage{
kl(mu1, cov1, mu2, cov2)
}
\arguments{
\item{mu1}{Mean vector}

\item{cov1}{Covariance matrix}

\item{mu2}{Mean vector}

\item{cov2}{Covariance matrix}
}
\description{
Divergence measures for multivariate normal distributions as used in the
diversityMatrix function.
}
