% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semtree-package.R
\docType{data}
\name{lgcm}
\alias{lgcm}
\title{Simulated Linear Latent Growth Curve Data}
\format{
\code{lgcm} is a matrix containing 400 rows and 8 columns of
simulated data. Longitudinal observations are o1-o5. Covariates are
agegroup, training, and noise.
}
\description{
This data set provides simple data to fit with a LGCM.
}
\author{
Andreas M. Brandmaier \email{brandmaier@mpib-berlin.mpg.de}
}
\keyword{datasets}
