% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeForests.R, R/semtree-package.R
\name{merge.semforest}
\alias{merge.semforest}
\title{Merge two SEM forests}
\usage{
\method{merge}{semforest}(x, y, ...)
}
\arguments{
\item{x}{A SEM Forest}

\item{y}{A second SEM Forest}

\item{\dots}{Extra arguments. Currently unused.}

\item{list()}{Extra arguments. Currently unused.}
}
\description{
This overrides generic base::merge() to merge two forests into one.

This overrides generic base::merge() to merge two forests into one.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.

Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\seealso{
\code{\link{semtree}}

\code{\link{semtree}}
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
