% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEstimates.R
\name{modelEstimates}
\alias{modelEstimates}
\title{Returns all estimates of a tree}
\usage{
modelEstimates(tree, ...)
}
\arguments{
\item{tree}{A semtree object.}

\item{\dots}{Optional arguments.}
}
\description{
Return model estimates of the tree.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
