\name{outliers}
\alias{outliers}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Find outliers based on case proximity
}
\description{
Compute outlier score based on proximity matrix.
}
\usage{
outliers(prox)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{prox}{
A proximity matrix.
}
}
\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier, John J. Prindle
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{proximity}}
}
