% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDependencePlot.R
\name{partialDependence}
\alias{partialDependence}
\alias{plot.partialDependence}
\alias{partialDependencePlot}
\title{Partial Dependence Plot}
\usage{
partialDependence(forest, reference.var, reference.param, support = NULL)
}
\arguments{
\item{forest}{A SEM forest}

\item{reference.var}{Label of the (independent) reference variable for which
partial dependence is plotted}

\item{reference.param}{Label of the (dependent) model parameter for which
partial dependence is plotted}

\item{support}{Number of grid points for interpolating the reference.var}
}
\description{
Partial dependence plot for the effect of an indepedent variable in the
dataset on the dependent outcome parameter selected. Returns a
partialDependence object that can be plotted via generic plot command.
}
\author{
Andreas M. Brandmaier
}
