% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParContrib.R
\name{plotParContrib}
\alias{plotParContrib}
\title{Plot a heatmap of the parameter contributions to the score-based test
statistics}
\usage{
plotParContrib(tree)
}
\arguments{
\item{tree}{a semtree object.}
}
\description{
Plots a heatmap with the standardized contributions of the model
 parameters to the score-based test statistics.
}
\author{
Manuel Arnold
}
