% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeStructure.R
\name{plotTreeStructure}
\alias{plotTreeStructure}
\title{Plot tree structure}
\usage{
plotTreeStructure(tree, type = 2, no.plot = FALSE, ...)
}
\arguments{
\item{tree}{a semtree object.}
}
\description{
Plots the the structure of a semtree object. This function is
similar to the \code{plot.semtree}, but is does not print the parameter
values in the leaf nodes and instead labels each node.
}
\author{
Manuel Arnold
}
