% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict.semforest}
\alias{predict.semforest}
\title{Predict method for semtree and semforest}
\usage{
\method{predict}{semforest}(object, data, type = "node_id", ...)
}
\arguments{
\item{object}{Object of class \code{semtree} or `semforest`.}

\item{data}{New test data of class \code{data.frame}. If no data is provided,
attempts to extract the data from the object.}

\item{type}{Type of prediction. One of `c('node_id')`. See Details.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Object of class \code{matrix}.
}
\description{
Predict method for semtree and semforest
}
\author{
Caspar J. van Lissa, Andreas Brandmaier
}
