% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.semtree.R
\name{predict.semtree}
\alias{predict.semtree}
\title{Predict terminal nodes}
\usage{
\method{predict}{semtree}(object, data, type = "node_id", ...)
}
\arguments{
\item{object}{semtree. A SEM tree node.}

\item{data}{a data.frame with observations used for prediction}

\item{type}{String. What type of prediction should be made? Currently, this
only supports 'node_id', which returns the id of a terminal node.}

\item{\dots}{Optional arguments passed to the function.}
}
\description{
This function takes a dataset and predicts the terminal node id that
matches the given observation
}
\author{
Caspar van Lissa, Andreas Brandmaier
}
