% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semforest.control.R
\name{semforest.control}
\alias{semforest.control}
\alias{print.semforest.control}
\title{SEM Forest Control Object}
\usage{
semforest.control(
  num.trees = 5,
  sampling = "subsample",
  control = NA,
  mtry = 2
)
}
\arguments{
\item{num.trees}{Number of trees.}

\item{sampling}{Sampling procedure. Can be subsample or bootstrap.}

\item{control}{A SEM Tree control object. Will be generated by default.}

\item{mtry}{Number of subsampled covariates at each node.}
}
\description{
A SEM Forest control object to tune parameters of the forest learning
algorithm.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
