% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semtree-package.R
\docType{data}
\name{semtree-package}
\alias{semtree-package}
\alias{.SCALE_METRIC}
\title{SEM Tree Package}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.SCALE_METRIC
}
\description{
SEM Tree Package
}
\keyword{datasets}
