% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{surrogate}
\alias{surrogate}
\title{use the current data 'mydata' and find a surrogate split}
\usage{
surrogate(rule, mydata, agreement_threshold = 0.7)
}
\arguments{
\item{rule}{a given split rule}

\item{mydata}{a dataset}

\item{agreement_threshold}{Numeric. Threshold for the agreement to consider a variable as surrogate}
}
\description{
use the current data 'mydata' and find a surrogate split
}
