% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{boruta}
\alias{boruta}
\alias{plot.boruta}
\alias{print.boruta}
\title{Run the Boruta algorithm on a sem tree}
\usage{
boruta(model, data, control = NULL, predictors = NULL, ...)
}
\arguments{
\item{model}{A template SEM. Same as in \code{semtree}.}

\item{data}{A dataframe to boruta on. Same as in \code{semtree}.}

\item{control}{A semforest control object to set forest parameters.}

\item{predictors}{An optional list of covariates. See semtree code example.}

\item{\dots}{Optional parameters.}

\item{constraints}{An optional list of covariates. See semtree code example.}
}
\value{
A boruta object.
}
\description{
Grows a series of SEM Forests following the boruta algorithm to determine
   feature importance as moderators of the underlying model.
}
\seealso{
\code{\link{semtree}} \code{\link{semforest}}
}
\author{
Priyanka Paul, Timothy R. Brick, Andreas Brandmaier
}
\keyword{models}
\keyword{multivariate}
\keyword{tree}
