% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.semtree}
\alias{coef.semtree}
\title{Return the parameter estimates of a given leaf of a SEM tree}
\usage{
\method{coef}{semtree}(object, ...)
}
\arguments{
\item{object}{semtree. A SEM tree node.}

\item{\dots}{Extra arguments. Currently unused.

 @exportS3Method coef semtree}
}
\description{
Return the parameter estimates of a given leaf of a SEM tree
}
