% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityMatrix.R
\name{diversityMatrix}
\alias{diversityMatrix}
\title{Diversity Matrix}
\usage{
diversityMatrix(forest, divergence = klsym, showProgressBar = TRUE)
}
\arguments{
\item{forest}{A SEM forest}

\item{divergence}{A divergence function such as hellinger or klsym}

\item{showProgressBar}{Boolean. Show a progress bar.}
}
\description{
Computes a diversity matrix using a distance function between trees
}
