\name{getDepth}
\alias{getDepth}
\title{
  Get the depth (or, height) a tree.
}
\description{
  Returns the length of the longest path from a root node to a leaf node.
}
\usage{
  getDepth(tree)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
    A \code{\link{semtree}} object
  }
}


\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier
}
