% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeight.R
\name{getHeight}
\alias{getHeight}
\title{Determine Height of a Tree}
\usage{
getHeight(tree)
}
\arguments{
\item{tree}{A SEM tree.}
}
\description{
Returns height of a SEM Tree, which equals to the length of the longest path
from root to a terminal node.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
