\name{getNodeById}
\alias{getNodeById}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get Node By Id
}
\description{
Return a node matching a given node ID
}
\usage{
getNodeById(tree, id)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
A SEM Tree object.
}
  \item{id}{
Numeric. A Node id.
}
}

\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier, John J. Prindle
}


