\name{getNumNodes}
\alias{getNumNodes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Tree Size
}
\description{
Counts the number of nodes in a tree.
}
\usage{
getNumNodes(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
A SEM tree object.
}
}

\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier, John J. Prindle
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
