\name{isLeaf}
\alias{isLeaf}
\title{
  Test whether a semtree object is a leaf.
}
\description{
  Tests whether a semtree object is a leaf. Returns TRUE or FALSE.
}
\usage{
  isLeaf(tree)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
    A \code{\link{semtree}} object
  }
}


\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier
}
