\name{lgcm}
\docType{data}
\alias{lgcm}
\title{Simulated Linear Latent Growth Curve Data}
\description{
  This data set provides simple data to fit with a LGCM.
}
\usage{

data(lgcm)

}

\format{
\code{lgcm} is a matrix containing 400 rows and 8 columns of simulated data. Longitudinal observations are o1-o5. Covariates are agegroup, training, and noise.

}

\author{
Andreas M. Brandmaier \email{brandmaier@mpib-berlin.mpg.de}}
\keyword{datasets}





