% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find outliers based on case proximity}
\usage{
outliers(prox)
}
\arguments{
\item{prox}{A proximity matrix.}
}
\description{
Compute outlier score based on proximity matrix.
}
\references{
Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger,
U. (2013). Structural equation model trees. \emph{Psychological Methods},
18(1), 71-86.
}
\seealso{
\code{\link{proximity}}
}
\author{
Andreas M. Brandmaier, John J. Prindle
}
