\name{semforest.control}
\alias{semforest.control}
\alias{print.semforest.control}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  SEM Forest Control Object
}
\description{
A SEM Forest control object to tune parameters of the forest learning algorithm.
}
\usage{
semforest.control(num.trees = 5, sampling = "subsample", control = NA, mtry = 2)
}
% print.semforest.control(x, ...)
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{num.trees}{
Number of trees.
}
  \item{sampling}{
Sampling procedure. Can be subsample or bootstrap.
}
  \item{control}{
A SEM Tree control object. Will be generated by default.
}
  \item{mtry}{
Number of subsampled covariates at each node.
}
}

\references{
  Brandmaier, A.M., Oertzen, T. v., McArdle, J.J., & Lindenberger, U. (2013). Structural equation model trees. \emph{Psychological Methods}, 18(1), 71-86.
}
\author{
  Andreas M. Brandmaier, John J. Prindle
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
