\name{toTable}
\alias{toTable}
\title{Tabular Representation of a SEM Tree}
\usage{
toTable(tree, added.param.cols=NULL, round.param=NULL) 
}
\description{
Converts a tree into a tabular representation. This may be useful as a textual representation for use in manuscripts.
}
\arguments{
\item{tree}{A SEM Tree object.}
\item{added.param.cols}{Add extra columns with parameter estimates.}
\item{round.param}{Number of digits to round parameter estiamtes}
}
\author{Andreas M. Brandmaier}

\references{

Brandmaier, A. M., Ram, N., Wagner, G. G., & Gerstorf, D. (in press). Terminal decline in well-being: The role of multi-indicator constellations of physical health and psychosocial correlates. \emph{Developmental Psychology}.
}